+++ test/videoIds.ts
@@ -1,0 +1,11 @@
+export const VIDEO_IDS = {
+    normal: 'dQw4w9WgXcQ',
+    private: 'wH31JWN3ziI',
+    beforePublication: 'fjbqNcKgm1Q',
+    limitedPublication: 'rSI0cB4cu1I',
+    forChildren: 'W0Ptl1DgiD8',
+    ageRestricted: 'HNeosKels-g',
+    unavailable: 'K4b9oHJk6q8',
+    invalid: 'invalid',
+    test: 'spLHYOElX6g',
+};